Incorporating review comments. Separate docker and model repo builds, add README, restructure repo
